deleting and editing comments

+++ app/Http/Livewire/CommentItem.php
@@ -9,6 +9,13 @@ class CommentItem extends Component
 {
     public Comment $comment;
 
+    public bool $editing = false;
+
+    protected $listeners = [
+        'cancelEditing' => 'cancelEditing',
+        'commentUpdated' => 'commentUpdated',
+    ];
+
     public function mount(Comment $comment)
     {
         $this->comment = $comment;
@@ -21,8 +28,33 @@ class CommentItem extends Component
 
     public function deleteComment()
     {
+        $user = auth()->user();
+        if (!$user) {
+            return $this->redirect('/login');
+        }
+
+        if ($this->comment->user_id != $user->id) {
+            return response('You are not allowed to perform this action', 403);
+        }
+
         $id = $this->comment->id;
+
         $this->comment->delete();
         $this->emitUp('commentDeleted', $id);
+    }
+
+    public function startCommentEdit()
+    {
+        $this->editing = true;
+    }
+
+    public function cancelEditing()
+    {
+        $this->editing = false;
+    }
+
+    public function commentUpdated()
+    {
+        $this->editing = false;
     }
 }
